dep upgrade and added timeline(Upcoming)
