+++ assets/styles/scss/sections/index/timeline.module.scss
@@ -1,0 +1,7 @@
+// timeline.module.scss
+
+.timeline {
+  position: relative;
+  width: 100%;
+  margin: 0 auto; // Center the timeline
+}
