+++ package.json
@@ -10,8 +10,8 @@
 		"build-webpack": "webpack --mode production"
 	},
 	"dependencies": {
-		"@fontsource/fira-code": "^5.0.8",
-		"@fontsource/inter": "^5.0.8",
+		"@fontsource/fira-code": "^5.0.12",
+		"@fontsource/inter": "^5.0.12",
 		"@fortawesome/fontawesome-svg-core": "^6.4.2",
 		"@fortawesome/free-brands-svg-icons": "6.4.2",
 		"@fortawesome/free-regular-svg-icons": "6.4.2",
@@ -95,7 +95,7 @@
 		"webpack-cli": "^5.1.4"
 	},
 	"name": "portfolio",
-	"version": "1.0.4",
+	"version": "1.0.5",
 	"description": "Open source portfolio website built with Next.js",
 	"main": "pages/index.tsx",
 	"repository": "https://github.com/muhammad-fiaz/portfolio.git",
